Revert "move vcc/cartridges into vcc/bus"
This reverts commit 85556042691f0ea350329d36d8257fc962910079.
