+++ libcommon/src/cartridges/capi_adapter_cartridge.cpp
@@ -15,12 +15,12 @@
 //	You should have received a copy of the GNU General Public License along with
 //	VCC (Virtual Color Computer). If not, see <http://www.gnu.org/licenses/>.
 ////////////////////////////////////////////////////////////////////////////////
-#include <vcc/bus/cartridges/capi_adapter_cartridge.h>
+#include <vcc/cartridges/capi_adapter_cartridge.h>
 #include <stdexcept>
 
 // C API hardware cartridge interface
 
-namespace vcc::bus::cartridges
+namespace vcc::cartridges
 {
 
 	namespace
